Add keyboard module
